refactor(server): rewire routes to module controllers (Phase 3-4)
auth.js, taskRoutes.js, categoryRoutes.js now import directly from
modules/ instead of the legacy controllers/ layer.
